backend/requirements.txt
flask
eth-brownie
psycopg2
python-dotenv
flask-cors